Simulated moult scores for 235 individuals with no heterogeneity in duration

A dataset containing the moult scores and capture dates for a simulated population of passerines including recaptures

recaptures

Format

A data frame with 299 rows and 13 variables:

id

individual identifier

sample_size

number of observations for the said individual

start_date

individual moult initiation date

duration

individual duration

set

simulation set

date_sampled

observation date (days since January 01)

pfmg_sampled

true PFMG at sampling occasion

pfmg_with_error

PFMG incorporating measurement error, if any

sampling_bias

descriptor of sampling bias type, if any

prop_recaptured_individuals

proportion of individuals with more than 1 capture

rel_tau_sd

relative population standard deviation of the moult duration

obs_error

measurement error

scenario

simulation scenario

Source

P. Boersch-Supan